optimize

+++ reno-kafka/reno-kafka-producer/src/main/java/cn/com/sherhom/reno/kafka/producer/entity/ProducerRunnerArgs.java
@@ -13,6 +13,6 @@ public class ProducerRunnerArgs {
     private int recordSize;
     private long num;
     //msg/sec
-    private int throughput;
+    private long throughput;
     private Stat stat;
 }
